frontend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
